requirements/base.txt
Django==1.10.4
djangorestframework==3.5.3
coreapi==2.1.1
openapi-codec==1.2.1
